move class code form file to spesific file
