+++ RestaurantApi.Dal/Services/MealsService.cs
@@ -24,20 +24,4 @@ namespace RestaurantApi.Dal.Services
 
         }
     }
-
-
-    public class OrdersService : IOrdersService
-    {
-        
-        public ICRUDService<Order> OrderService { get; set; }
-        public IOrderItemCRUDService OrderItemService { get; set; }
-
-        public OrdersService(ICRUDService<Order> orderService,
-                            IOrderItemCRUDService orderItemService
-            )
-        {
-            OrderService = orderService;
-            OrderItemService = orderItemService;
-        }
-    }
 }
